+++ src/CR-Tools/implement/Data/LOFAR_StationGroup.cc
@@ -503,20 +503,20 @@ namespace DAL { // Namespace DAL -- begin
 
   // ------------------------------------------------------------------------- fx
 
-  casa::Matrix<short> LOFAR_StationGroup::fx (int const &start,
-					      int const &nofSamples)
+  casa::Matrix<double> LOFAR_StationGroup::fx (int const &start,
+					       int const &nofSamples)
   {
     /* Check minimal condition for operations below. */
     if (groupID_p < 1) {
-      return casa::Matrix<short> (1,1,0);
+      return casa::Matrix<double> (1,1,0);
     }
 
     uint nofDipoles = nofDipoleDatasets();
-    casa::Matrix<short> data (nofSamples,nofDipoles);
+    casa::Matrix<double> data (nofSamples,nofDipoles);
 
     /* Go through the set of dipole datasets and retrieve the data */
 
-    casa::Vector<short> tmp (nofSamples);
+    casa::Vector<double> tmp (nofSamples);
 
     for (uint n(0); n<nofDipoles; n++) {
       // get the channel data ...
@@ -530,29 +530,29 @@ namespace DAL { // Namespace DAL -- begin
 
   // ------------------------------------------------------------------------- fx
   
-  casa::Matrix<short> LOFAR_StationGroup::fx (int const &start,
-					      int const &nofSamples,
-					      std::vector<uint> const &dipoleSelection)
+  casa::Matrix<double> LOFAR_StationGroup::fx (int const &start,
+					       int const &nofSamples,
+					       std::vector<uint> const &dipoleSelection)
   {
     /* Check minimal condition for operations below. */
     if (groupID_p < 1) {
-      return casa::Matrix<short> (1,1,0);
+      return casa::Matrix<double> (1,1,0);
     }
 
     uint nofDipoles = dipoleSelection.size();
-    casa::Matrix<short> data (nofSamples,nofDipoles);
+    casa::Matrix<double> data (nofSamples,nofDipoles);
 
     /* Go through the set of dipole datasets and retrieve the data */
-
-    casa::Vector<short> tmp (nofSamples);
-
+    
+    casa::Vector<double> tmp (nofSamples);
+    
     for (uint n(0); n<nofDipoles; n++) {
       // get the channel data ...
       tmp = datasets_p[dipoleSelection[n]].fx(start,nofSamples);
       // ... and add them to the returned array
       data.column(n) = tmp;
     }
-
+    
     return data;
   }
